Add support for Server-Sice RPC call. Move MQTT client related code into MqttGatewayClient.java
